Changed floats to ints in all but main.cpp;

+++ partnerProject/Vec3.h
@@ -4,14 +4,14 @@ public:
 	//Declares variables
 	float x, y, z;
 	//Default constructor
-	Vec3(const float xIn = 0.0f, const float yIn = 0.0f, const float zIn = 0.0f) {
+	Vec3(const int xIn = 0, const float yIn = 0, const float zIn = 0) {
 		x = xIn;
 		y = yIn;
 		z = zIn;
 	}
 	//Returns the dot product between two Vec3s
-	float Dot(const Vec3& vec3In) const;
+	int Dot(const Vec3& vec3In) const;
 	//Returns the magnitude of the Vec3
-	float Mag() const;
+	int Mag() const;
 };
